+++ packages/components/CHANGELOG.md
@@ -14,6 +14,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### [Unreleased]
 
+#### Added 
+
+- Partials: `getMetaTagsAndIconLinks`, `getComponentChunkLinks`, `getIconLinks` and `getFontLinks` support new format option `js`
+  ([#3179](https://github.com/porsche-design-system/porsche-design-system/pull/3179))
+
 ### [3.15.0-rc.0] - 2024-04-05
 
 #### Changed